`FrameRef` for all types of frames
